refactor: separate different apis
