+++ hstream-admin-server/hstream-admin-server.cabal
@@ -27,14 +27,28 @@ executable hstream-admin-server
     , servant
     , servant-server
     , warp    
-    , hstream
+    , hstream-admin-server
     , hstream-store
     , hstream-processing
+    , Z-Data
+
+
+  hs-source-dirs:   app
+  default-language: Haskell2010
+
+library
+  hs-source-dirs:   src
+  exposed-modules:
+    HStream.Admin.Server.Stream
+  build-depends:
+    , base              >=4.11 && <5
+    , servant
+    , servant-server
+    , warp    
+    , hstream
+    , hstream-store
     , hstream-connector
     , aeson
     , Z-Data
     , containers
     , text
-
-  hs-source-dirs:   app
-  default-language: Haskell2010
